Changed type to qbyte to avoid conflict with std
The code defines a byte with typedef which creates compile errors on
gcc-12 (and probably other new compilers as well) due to conflicts
between the defined type byte and the std::byte that is pulled in due
to using namespace std; everywhere. This commit changes the byte type to
qbyte and therefore avoids these conflicts.